rm root.go
