--- cmd/csuf/root.go
@@ -1,36 +1,0 @@
-package main
-
-import (
-	"log"
-	"os"
-
-	"github.com/acmcsufoss/api.acmcsuf.com/internal/cli/announcements"
-	"github.com/acmcsufoss/api.acmcsuf.com/internal/cli/events"
-
-	"github.com/spf13/cobra"
-)
-
-var Version = "dev"
-
-var rootCmd = &cobra.Command{
-	Use:     "csuf",
-	Short:   "A CLI tool to help manage the API of the CSUF ACM website",
-	Version: Version,
-}
-
-func Execute() {
-
-	// Logging the error, prefix is date, time, and what file the log is from
-	log.SetFlags(log.Ldate | log.Ltime | log.Lshortfile)
-
-	if err := rootCmd.Execute(); err != nil {
-		log.Println("Error with CLI:", err)
-		os.Exit(1)
-	}
-
-}
-
-func init() {
-	rootCmd.AddCommand(events.CLIEvents)
-	rootCmd.AddCommand(announcements.CLIAnnouncements)
-}
